backend/pyproject.toml
 [project]
 name = "agentic-med-assistant"
 version = "0.1.0"
 description = "Agentic multimodal RAG assistant orchestrated with LangGraph"
 readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
   "fastapi>=0.115.0",
   "uvicorn[standard]>=0.30.0",
   "pydantic>=2.7.0",
   "pydantic-settings>=2.4.0",
   "structlog>=24.1.0",
   "prometheus-client>=0.21.0",
   "langchain>=0.3.0",
   "langgraph>=0.2.0",
   "qdrant-client>=1.9.0",
   "docling>=2.0.0",
   "unstructured>=0.15.0",
   "transformers>=4.44.0",
   "torch>=2.3.0",
   "Pillow>=10.3.0",
   "sentence-transformers>=3.0.0",
   "faster-whisper>=1.0.0",
   "piper-tts>=1.2.0",
   "python-multipart>=0.0.9",
   "orjson>=3.10.0",
   "httpx>=0.27.0",
   "google-generativeai>=0.5.2",
   "langchain-google-genai>=1.0.4",
   "tenacity>=8.2.0",
   "aiofiles>=23.2.0",
 ]

 [project.optional-dependencies]
 dev = [
   "pytest>=8.1.0",
   "pytest-asyncio>=0.23.0",
   "httpx>=0.27.0",
   "mypy>=1.10.0",
   "ruff>=0.5.0",
   "black>=24.4.0",
   "types-orjson>=3.6.2",
 ]

 [build-system]
 requires = ["setuptools>=68", "wheel"]
 build-backend = "setuptools.build_meta"
